py-sender/requirements.txt
msgpack
pyzmq
Pillow
numpy